requirements.txt
paramiko~=3.3.1
pandas~=2.1.0
python-nmap~=0.7.1
network~=0.1
pyinstaller==5.13.2
Pillow==10.0.1
BeautifulSoup4==4.12.2
requests==2.31.0
webdriver-manager==4.0.0
selenium==4.12.0
pytube~=15.0.0
pyperclip~=1.8.2
wakeonlan~=3.0.0
customtkinter~=5.2.0
moviepy~=1.0.3
